SubjectDetailActivity navigateUp() method changed, in order to de-hardcode minimum birthdate year
